+++ src/components/DailyReport/share_button/tweet_button.tsx
@@ -1,0 +1,24 @@
+import React, { useState } from 'react';
+import { Button } from '@material-ui/core';
+
+type TweetButtonProps = {
+  text: string;
+};
+
+export const TweetButton: React.FC<TweetButtonProps> = (props: TweetButtonProps) => {
+  const tweet = `https://twitter.com/intent/tweet?text=${encodeURIComponent(props.text)}`;
+  return (
+    <Button
+      style={{
+        margin: '5px',
+        textTransform: 'none',
+        float: 'right',
+      }}
+      variant="contained"
+      color="primary"
+      href={tweet}
+    >
+      Tweetする
+    </Button>
+  );
+};
